feat: added hook for looking up users with server action
